Remove Null Data from Publisher
Monasca has added a check for null data. When the publisher does not
find the meter value or metadata it defaults it to Null.

This causes Monasca API to reject the sample. This is particularly
evident in the Central and Compute Agents since they collect data
that may be null.

Change-Id: I51ca07ec10c6b414b6a26c057edc09487053f982